fixes to compile linux; fixed a couple of memory leaks!

+++ tests/Parser53TestClass.cpp
@@ -1359,7 +1359,7 @@ TEST_FIXTURE(Parser53TestClass, ExpressionObserverWithClosure) {
 	CHECK_VARIABLE("$b", closure->Parameters[1]);
 	CHECK_VECTOR_SIZE(1, closure->LexicalVars);
 	CHECK_VARIABLE("$c", closure->LexicalVars[0]);
-	CHECK_EQUAL(1, closure->Statements.Size());
+	CHECK_EQUAL((size_t)1, closure->Statements.Size());
 
 	CHECK_EQUAL(pelet::StatementClass::EXPRESSION, closure->Statements.TypeAt(0));
 	pelet::StatementClass* stmt = closure->Statements.At(0);
